+++ CHANGELOG.md
@@ -5,3 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+# [0.3.0] - 2021-03-15
+### Changed
+- Pass closures to wrapped signals correctly
+- Return wrapped proxy as `<Type>Proxy` instead of trait object `Proxy`
+- Enable more signals
+- Add basic signal example